+++ library/src/main/java/com/davidstemmer/screenplay/scene/transformer/NoAnimationTransformer.java
@@ -1,14 +1,19 @@
-package com.davidstemmer.screenplay.sample.simple.scene.transformer;
+package com.davidstemmer.screenplay.scene.transformer;
+
+/**
+ * Created by weefbellington on 4/2/15.
+ */
 
 import com.davidstemmer.screenplay.SceneCut;
 import com.davidstemmer.screenplay.flow.Screenplay;
 import com.davidstemmer.screenplay.scene.Scene;
 
-/**
- * Created by weefbellington on 10/17/14.
- */
+import javax.inject.Inject;
+
+
 public class NoAnimationTransformer implements Scene.Transformer {
 
+    @Inject
     public NoAnimationTransformer() {}
 
     @Override