+++ sample-mortar/src/main/java/com/davidstemmer/screenplay/sample/mortar/module/ApplicationComponent.java
@@ -2,10 +2,11 @@ package com.davidstemmer.screenplay.sample.mortar.module;
 
 import android.app.Application;
 
-import com.davidstemmer.screenplay.SimpleActivityDirector;
+import com.davidstemmer.screenplay.MutableStage;
 import com.davidstemmer.screenplay.flow.Screenplay;
 import com.davidstemmer.screenplay.sample.mortar.presenter.DrawerPresenter;
 import com.davidstemmer.screenplay.sample.mortar.presenter.NavigationMenuPresenter;
+import com.davidstemmer.screenplay.sample.mortar.scene.WelcomeScene;
 import com.davidstemmer.screenplay.sample.mortar.scene.transformer.ActionDrawerTransformer;
 
 import javax.inject.Singleton;
@@ -19,16 +20,12 @@ import flow.Flow;
 @Singleton // Constraints this component to one-per-application or unscoped bindings.
 @Component(modules = ApplicationModule.class)
 public interface ApplicationComponent {
-    // application class
     Application application();
-    // screenplay boilerplate
-    SimpleActivityDirector director();
     Flow flow();
     Screenplay screenplay();
-    // screenplay transformers
     ActionDrawerTransformer actionDrawerTransformer();
-    // presenter classes
+    MutableStage stage();
     DrawerPresenter drawerPresenter();
     NavigationMenuPresenter menuPresenter();
-
+    WelcomeScene initialScene();
 }