+++ sample-mortar/src/main/java/com/davidstemmer/screenplay/sample/mortar/module/ApplicationModule.java
@@ -2,17 +2,10 @@ package com.davidstemmer.screenplay.sample.mortar.module;
 
 import android.app.Application;
 
-import com.davidstemmer.screenplay.SimpleActivityDirector;
+import com.davidstemmer.screenplay.MutableStage;
 import com.davidstemmer.screenplay.flow.Screenplay;
-import com.davidstemmer.screenplay.sample.mortar.MainActivity;
-import com.davidstemmer.screenplay.sample.mortar.presenter.NavigationMenuPresenter;
-import com.davidstemmer.screenplay.sample.mortar.scene.DialogScene;
-import com.davidstemmer.screenplay.sample.mortar.scene.PagedScene1;
-import com.davidstemmer.screenplay.sample.mortar.scene.PagedScene2;
-import com.davidstemmer.screenplay.sample.mortar.scene.PagedScene3;
-import com.davidstemmer.screenplay.sample.mortar.scene.StackedScene;
-import com.davidstemmer.screenplay.sample.mortar.scene.WelcomeScene;
-import com.davidstemmer.screenplay.sample.mortar.scene.transformer.HorizontalSlideTransformer;
+import com.davidstemmer.screenplay.scene.StandardScene;
+
 import javax.inject.Singleton;
 
 import dagger.Module;
@@ -23,21 +16,7 @@ import flow.Flow;
 /**
  * Created by weefbellington on 10/2/14.
  */
-@Module(injects = {
-        HorizontalSlideTransformer.class,
-        DialogScene.class,
-        DialogScene.Presenter.class,
-        StackedScene.class,
-        StackedScene.Presenter.class,
-        MainActivity.class,
-        NavigationMenuPresenter.class,
-        PagedScene1.class,
-        PagedScene1.Presenter.class,
-        PagedScene2.class,
-        PagedScene2.Presenter.class,
-        PagedScene3.class,
-        WelcomeScene.class,
-})
+@Module
 public class ApplicationModule {
 
     private final Application application;
@@ -46,24 +25,30 @@ public class ApplicationModule {
         this.application = application;
     }
 
-    @Provides
+    @Provides @Singleton
     Application provideApplication() {
         return application;
     }
 
     @Provides @Singleton
-    SimpleActivityDirector provideActivityDirector() {
-        return new SimpleActivityDirector();
+    Flow provideFlow(Screenplay screenplay) {
+        return new Flow(Backstack.single(new StandardScene() {
+            @Override
+            public Transformer getTransformer() {
+                return null;
+            }
+        }), screenplay);
     }
 
     @Provides @Singleton
-    Screenplay provideScreenplay(SimpleActivityDirector director) {
-        return new Screenplay(director);
+    Screenplay screenplay() {
+        return new Screenplay();
     }
 
+
     @Provides @Singleton
-    Flow provideFlow(WelcomeScene welcomeScene, Screenplay screenplay) {
-        return new Flow(Backstack.single(welcomeScene), screenplay);
+    MutableStage provideStage() {
+        return new MutableStage();
     }
 
 }