+++ sample-mortar/src/main/java/com/davidstemmer/screenplay/sample/mortar/scene/transformer/HorizontalSlideTransformer.java
@@ -12,7 +12,6 @@ import javax.inject.Singleton;
  * Created by weefbellington on 10/2/14.
  */
 
-@Singleton
 public class HorizontalSlideTransformer extends TweenTransformer {
 
     private static final Params params = new Params();