+++ sample-mortar/src/main/java/com/davidstemmer/screenplay/sample/mortar/scene/transformer/VerticalSlideTransformer.java
@@ -6,13 +6,10 @@ import com.davidstemmer.screenplay.sample.mortar.R;
 import com.davidstemmer.screenplay.scene.transformer.TweenTransformer;
 
 import javax.inject.Inject;
-import javax.inject.Singleton;
 
 /**
  * Created by weefbellington on 10/2/14.
  */
-
-@Singleton
 public class VerticalSlideTransformer extends TweenTransformer {
 
     private static final Params params = new Params();